Source commit: 2c3cdff2c1af637c9ccf0aa5dee28772eb42cc97

+++ libs/neon/test/common/child.h
@@ -118,7 +118,7 @@ struct serve_file_args {
 int serve_file(ne_socket *sock, void *ud);
 
 /* set to value of C-L header by discard_request. */
-extern int clength;
+extern unsigned long clength;
 
 /* Sleep for a short time. */
 void minisleep(void);
